all basic functionalities done

+++ frontend/src/pages/UserDashboard/UserDashboard.jsx
@@ -6,9 +6,13 @@ import APICard from "../../components/APICard/APICard";
 import { userContext } from "../../userContext";
 import Alert from "../../utils/Alert/Alert";
 import NewAPI from "../../components/Navbar/NewAPI/NewAPI";
+import Pagination from "../../utils/Pagination/Pagination";
 const UserDashboard = () => {
   const userData = React.useContext(userContext);
+  const [currPage, setCurrPage] = useState(1);
+  const [apisPerPage, setApisPerPage] = useState(3);
   const [showModal, setShowModal] = useState(false);
+  const [loading, setLoading] = useState(false);
   const [apis, setAPIs] = useState([]);
   const [updateID, setUpdateID] = useState("");
   const [message, setMessage] = useState({
@@ -16,26 +20,39 @@ const UserDashboard = () => {
     success: true,
     message: "",
   });
+  const clickOutside = (e) => {
+    const modal = document.getElementById("new-api-modal");
+    if (e.target == modal) {
+      setShowModal(false);
+    }
+  };
+  useEffect(() => {
+    window.addEventListener("click", clickOutside);
+  }, []);
   const getAPIs = async () => {
     try {
       let userId = userData.user.id;
       let res = await axios.get(`/api/userAPIs/apis/user/${userId}`);
       setAPIs(res.data.apis);
-      console.log(res.data.apis);
     } catch (error) {
-      console.log(error.response.data);
+      console.log(error);
     }
+    setLoading(false);
   };
   useEffect(() => {
+    setLoading(true);
     getAPIs();
   }, []);
   const handleDelete = async (id) => {
+    if (!confirm("Are you sure, you want to delete this API?")) return;
     console.log("Key is " + id);
     try {
       const res = await axios.delete(`/api/userAPIs/apis/${id}`);
       console.log(res);
       setMessage({ display: true, success: true, message: res.data.message });
-      getAPIs();
+      setTimeout(() => {
+        history.go(0);
+      }, 1000);
     } catch (error) {
       console.log(error.response.data);
       setMessage({
@@ -51,39 +68,69 @@ const UserDashboard = () => {
     setUpdateID(id);
     setShowModal(true);
   };
+  const indexOfLastJob = currPage * apisPerPage;
+  const indexOfFirstJob = indexOfLastJob - apisPerPage;
+  let currAPIs = [];
+  //Get current Jobs to be displayed on that particular page
+  if (indexOfFirstJob >= apis.length) {
+    currAPIs = apis;
+  } else {
+    currAPIs = apis.slice(indexOfFirstJob, indexOfLastJob);
+  }
+
   return (
     <div>
       {showModal && (
         <NewAPI
+          getAPIs={getAPIs}
           updateData={{ isUpdate: true, id: updateID }}
           setShowModal={setShowModal}
         />
       )}
-      {apis.length == 0 ? (
+      {loading ? (
         <div id="preloader"></div>
       ) : (
         <>
-          <div id="alert" className={styles["alert-box"]}>
-            <Alert
-              setMessage={setMessage}
-              display={message.display}
-              success={message.success}
-              message={message.message}
-            />
-          </div>
-          <div className={styles["apis"]}>
-            {apis.map((api, index) => (
-                <APICard
-                  key={index}
-                  id={api._id}
-                  apiImage={api.imageUrl || "/api_logo.png"}
-                  apiTitle={api.name}
-                  apiDescription={api.description || "This is a very cool API"}
-                  handleDelete={handleDelete}
-                  handleEdit={handleEdit}
+          {apis != null && apis.length === 0 ? (
+            <p>There are no APIS....</p>
+          ) : (
+            <>
+              <div id="alert" className={styles["alert-box"]}>
+                <Alert
+                  setMessage={setMessage}
+                  display={message.display}
+                  success={message.success}
+                  message={message.message}
                 />
-            ))}
-          </div>
+              </div>
+              <div className={styles["apis"]}>
+                {currAPIs.map((api, index) => (
+                  <APICard
+                    key={index}
+                    api={{
+                      apiID: api._id,
+                      apiImage: api.imageUrl || "/api_logo.png",
+                      apiTitle: api.name,
+                      apiDescription:
+                        api.description || "This is a very cool API",
+                      apiEndpoints: api.endpoints,
+                    }}
+                    handleDelete={handleDelete}
+                    handleEdit={handleEdit}
+                  />
+                ))}
+              </div>
+              {apis.length > apisPerPage && (
+                <div className={styles["pagination"]}>
+                  <Pagination
+                    count={Math.ceil(apis.length / apisPerPage)}
+                    currPage={currPage}
+                    setCurrPage={setCurrPage}
+                  />
+                </div>
+              )}
+            </>
+          )}
         </>
       )}
     </div>
